requirements.txt
Flask==2.3.2
requests==2.28.2
joblib==1.2.0
dvc==3.0.0
remla24-team8-lib-ml == 0.1.1